split out num into separate features
