+++ .github/workflows/ci.yml
@@ -93,6 +93,9 @@ jobs:
           - os: ubuntu-latest
             rust: stable
             recipe: check-tests-all-features
+          - os: ubuntu-latest
+            rust: stable
+            recipe: check-num-subfeatures
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
         with:
